refactor | afc13 : move AlgoMakerVolume to volume mod
