+++ crates/sal-3dlib-core/src/ops/boolean/volume.rs
@@ -19,3 +19,20 @@ pub struct VolumeConf {
     /// so all fields of [OpConf] also make sense.
     pub op_conf: OpConf,
 }
+///
+/// Algorithm to volume together three sets of the objects.
+pub trait AlgoMakerVolume<A, B, C> {
+    type Error;
+    //
+    //
+    fn build<'a>(
+        a: impl IntoIterator<Item = &'a A>,
+        b: impl IntoIterator<Item = &'a B>,
+        c: impl IntoIterator<Item = &'a C>,
+    ) -> Result<Self, Self::Error>
+    where
+        Self: Sized,
+        A: 'a,
+        B: 'a,
+        C: 'a;
+}
